Chown plugins to root

+++ installation/raspiBackupInstallUI.sh
@@ -1905,6 +1905,11 @@ function extensions_install_execute() {
 		return
 	fi
 
+	if ! chown root.root $FILE_TO_INSTALL_ABS_PATH/${RASPIBACKUP_NAME}_*.sh &>>"$LOG_FILE"; then
+		unrecoverableError $MSG_SAMPLEEXTENSION_INSTALL_FAILED "chmod extensions"
+		return
+	fi
+
 	if ! rm -f "$SAMPLEEXTENSION_TAR_FILE" 2>>"$LOG_FILE"; then
 		unrecoverableError $MSG_UNINSTALL_FAILED "$SAMPLEEXTENSION_TAR_FILE"
 		return
